refactor: rooms use cases

+++ backend/src/useCases/DeleteRoomUseCase.ts
@@ -1,3 +1,4 @@
+import { AppError } from "../errors/AppError";
 import { IRoomsRepository } from "../repositories/IRoomsRepository";
 
 type DeleteRoomUseCaseDTO = {
@@ -12,6 +13,11 @@ export class DeleteRoomUseCase {
   }
 
   async execute(data: DeleteRoomUseCaseDTO) {
+    const roomExists = await this.roomsRepository.findById({ id: data.id })
+    if (!roomExists) {
+      throw new AppError('Sala não encontrada', 404)
+    }
+
     await this.roomsRepository.delete(data)
   }
 }
